rename db struct name
